ENH: Added OpenFiles/OpenDocuments event handling on Mac
by replacing the Qt event handlers for the Carbon kAEOpenDocuments event
and the openFiles method of the QCocoApplicationDelegate class. The
latter is a bit of a hack because I use the Objective-C runtime to
inject my own method into QCocoApplicationDelegate.

The whole thing is necessary because Qt splits a single Apple-event with
multiple files into multiple QFileOpenEvent with a single file each.
This way it is impossible for the user to open time-sequences. The class
replacing QFileOpenEvent is pqOpenFilesEvent.

Signed-off-by: Michael Wild <themiwi@users.sourceforge.net>